+++ CHANGELOG.md
@@ -5,10 +5,16 @@ PrestaShop 1.6.x, 1.7.x and 8.x
 
 # Changelog
 
-## [3.7.4]
+## [3.7.5]
 ### Added
 - Add support for Trustly.
 - Create order on fail callback if reserve amount is greater than zero.
+
+## [3.7.4]
+### Fixed
+- Fix: Improve exception handling and error logging.
+- Fix: Round off unit price in order line to 3 decimal digits.
+- Fix: Do not change order status on capture when triggered by the order hook for status shipped or delivered.
 
 ## [3.7.3]
 ### Fixed